requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
python-telegram-bot==20.0
